Python bindings now "pygetdata".
Added the "Barth-style" meta field specification:
parent/child <TYPE> <blah, blah, blah>

Completion of the bindings tests for F77, F95, C++, and Python, plus other miscellaneous tests of the C API.

Fixed bugs found thereby:
* Multiply user-supplied frame limits by spf in get_frame_subset64()
* Fix error strings for GD_E_DELETE.
* Properly handle repr in get_constant()
* Properly handle byte swapping complex data in putdata()
* Remove spurious free in putdata()
* Fix E->table to properly contain the "external name".
* Close linterp tables after reading them.
* Allow INDEX as a metafield name.
* Recompute comp_scal upon modification.
* Fixed accidental failure of dirfile_madd_string().
* [PYTHON] Fix initialisation of PHASE entries.
* [PYTHON] Make parameter list for strings optional
* [PYTHON] Fix various incorrect field type checks for entry attributes
* [C++] Properly handling renaming metafields
* [C++] Fix C API call in Fragment::SetEncoding().
* [C++] Add many empty constructors.
* [C++] Add a method to dissociate entry objects from their dirfile.
* [C++] Include all headers in dirfile.h
* [F77] Don't crash on a bad dirfile number.
* [F77] Return empty string instead of NULL if the C API returns "".
* [F77] Report string lengths to be 1 less in length than their C counterparts.
* [F77] Use gd_entry%poly_ord instead of gd_entry%n_fields where appropriate.
* [F95] Fix calls in fget_nmfields(), fget_mvector_list(), and for PHASE in fget_entry().
* Added the missing bindings:
  - Python: dirfile_alter, get_fragment_index, get_raw_filename, dirfile_move, dirfile_rename
  - C++: Entry::SetName Entry::SetFragmentIndex Dirfile::FragmentIndex Dirfile::Name Entry::Name Entry::ComplexScalars Entry::ConstType

+++ src/types.c
@@ -697,7 +697,7 @@ void _GD_ConvertType(DIRFILE* D, const void *data_in, gd_type_t in_type,
           return;
         case GD_FLOAT64:
           for (i = 0; i < n; i++)
-            ((float*)data_out)[i] = (float)((double complex*)data_in)[i];
+            ((double*)data_out)[i] = (double)((double complex*)data_in)[i];
           return;
         case GD_COMPLEX64:
           for (i = 0; i < n; i++)
